% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix.R
\name{prefix}
\alias{prefix}
\title{Create a prefix function to shorten the use path}
\usage{
prefix(prefix_path)
}
\arguments{
\item{prefix_path}{a module path that defines the prefix.}
}
\value{
a prefix function that can be further used in use and useonly.
}
\description{
prefix() provides a useful mechanism to shorten an import path
verbosity and extract a common prefix once and for all.
The returned entity can be used as is, or it can be called as a function.
calling the function with a path will "extend" the path and return a new
prefix function with the extended path.
}
\examples{
\dontrun{
  mylongmoduledir <- prefix("/my/long/module/dir")
  mysymbol <- useonly(mylongmoduledir, "mysymbol")
  mysymbol <- useonly(mylongmoduledir(), "mysymbol")

  mylong <- prefix("/my/long")
  mylongmoduledir <- mylong("module/dir")
  mysymbol <- useonly(mylongmoduledir, "mysymbol")
}

}
