% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_path.R
\name{prepend_search_path}
\alias{prepend_search_path}
\title{Adds a given path to the list of looked up directories}
\usage{
prepend_search_path(path)
}
\arguments{
\item{path}{a path on the disk where to find modules.}
}
\description{
Adds a given path to the list of searched paths when looking
for a given module. The path is prepended is prepended, which means it is
at a higher priority than the current list. It can be called multiple times.
}
\examples{
\dontrun{
  prepend_search_path("dir/where/my/modules/are")
}


}
