% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useonly.R
\name{useonly}
\alias{useonly}
\title{Import a module and return only the specified symbol}
\usage{
useonly(path, symbol_name)
}
\arguments{
\item{path}{a path, or a function returning a path}

\item{symbol_name}{the name of the symbol to import}
}
\value{
the symbol
}
\description{
Imports a module, but return only the specified symbol
(function or variable), instead of the whole module.
Equivalent in python to `from module import function`
}
\examples{
\dontrun{
  myfunc <- useonly("/my/mod/submod", "myfunc")
  myfunc()
}

}
